requirements.txt
absl-py==2.3.1
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.14
ml_dtypes==0.5.1
narwhals==1.48.1
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==4.25.8
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sounddevice==0.5.2
streamlit==1.47.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
